requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
dataclasses-json==0.5.9
docarray==0.35.0
frozenlist==1.3.3
google-search-results==2.4.2
idna==3.4
langchain==0.0.229
langchainplus-sdk==0.0.20
markdown-it-py==3.0.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.2
packaging==23.1
pydantic==1.10.11
Pygments==2.15.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
SQLAlchemy==2.0.18
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
